// legacy_mobile_ionic/src/app/models/db/i-offline-document.ts
import {DocumentSyncStatus, DocumentTypeAcronyms} from "../../common/enum";

/**
 * Represent the model of the offline document entity
 */
export interface IOfflineDocument
{
    /**
     * The serialized document
     */
    RawDocument: string;
    /**
     * Indicates if should send the document
     */
    ShouldSend: number;
    /**
     * The local record id
     */
    Id: number;
    /**
     * The type of the serialized document (Type of DocumentTypeAcronyms)
     */
    DocumentType: string;
    /**
     * The total of the serialized document
     */
    DocumentTotal: number;
    /**
     * Document key generated by the application
     */
    DocumentKey: string;
    /**
     * Represent the status of the document
     */
    TransactionStatus: DocumentSyncStatus;
    /**
     * Contains the result of trying to synchronize the document
     */
    TransactionDetail: string;
    /**
     * Date when the document was created offline
     */
    OfflineDate: string;
    /**
     * The id of the user assign that create this document
     */
    UserAssignId: number;
    /**
     * The type of the transaction
     */
    TransactionType: string;
}

/**
 * Represent a model used to send the documents to the remote database and manage the result
 */
export interface IOfflineDocumentHandler 
{
    /**
     * The offline document information
     */
    OfflineDocument: IOfflineDocument;
    /**
     * Indicates if the document was synchronized
     */
    Finished: boolean;
    /**
     * Indicates if the document was successful synchronized
     */
    Result: boolean;
    /**
     * The information of the document
     */
    Info: string;
    /**
     * Represent the status of the document
     */
    TransactionStatus: number;
    /**
     * Contains the result of trying to synchronize the document
     */
    TransactionDetail: string;
}